Move post status registration to cpt file
